Split command sidebar surface tests
