+++ Tests/QuillCodeAppTests/WorkspaceSurfaceTests.swift
@@ -218,241 +218,6 @@ final class WorkspaceSurfaceTests: XCTestCase {
         XCTAssertEqual(model.surface().commands.first { $0.id == "disconnect-all" }?.isEnabled, false)
     }
 
-    func testSidebarBulkSelectionArchivesAndDeletesChats() {
-        let project = ProjectRef(name: "QuillCode", path: "/tmp/QuillCode")
-        let first = ChatThread(title: "Run whoami", projectID: project.id)
-        let second = ChatThread(title: "Check diff", projectID: project.id)
-        let fallback = ChatThread(title: "Review tests", projectID: project.id)
-        let model = QuillCodeWorkspaceModel(root: QuillCodeRootState(
-            projects: [project],
-            selectedProjectID: project.id,
-            threads: [first, second, fallback],
-            selectedThreadID: first.id
-        ))
-
-        model.startSidebarSelection(selecting: first.id)
-        model.toggleSidebarThreadSelection(second.id)
-
-        var surface = model.surface()
-        XCTAssertTrue(surface.sidebar.isSelectionMode)
-        XCTAssertEqual(surface.sidebar.selectionLabel, "2 chats selected")
-        XCTAssertEqual(Set(surface.sidebar.items.filter(\.isBulkSelected).map(\.id)), [first.id, second.id])
-        XCTAssertEqual(surface.sidebar.bulkActions.first { $0.kind == .archive }?.isEnabled, true)
-
-        XCTAssertTrue(model.performSidebarBulkAction(.archive))
-        surface = model.surface()
-
-        XCTAssertFalse(surface.sidebar.isSelectionMode)
-        XCTAssertEqual(Set(surface.sidebar.archivedItems.map(\.id)), [first.id, second.id])
-        XCTAssertEqual(surface.sidebar.selectedThreadID, fallback.id)
-
-        model.selectAllSidebarThreads()
-        surface = model.surface()
-        XCTAssertEqual(surface.sidebar.selectionLabel, "3 chats selected")
-        XCTAssertTrue(model.performSidebarBulkAction(.delete))
-
-        surface = model.surface()
-        XCTAssertEqual(surface.sidebar.items.count, 0)
-        XCTAssertNil(surface.sidebar.selectedThreadID)
-        XCTAssertFalse(surface.sidebar.isSelectionMode)
-    }
-
-    func testSidebarSearchExcludesHiddenToolFeedback() {
-        let thread = ChatThread(title: "Visible thread", messages: [
-            .init(role: .user, content: "run whoami"),
-            .init(role: .tool, content: #"{"result":"secret internal feedback"}"#),
-            .init(role: .assistant, content: "Output:\nquill")
-        ])
-
-        let item = SidebarItem(thread: thread)
-        let sidebar = SidebarSurface(
-            items: [SidebarItemSurface(item: item, selectedThreadID: thread.id)],
-            selectedThreadID: thread.id
-        )
-
-        XCTAssertEqual(sidebar.filteredItems(matching: "secret internal feedback"), [])
-        XCTAssertEqual(sidebar.filteredItems(matching: "whoami").map(\.id), [thread.id])
-    }
-
-    func testComposerShowsFilteredSlashSuggestions() {
-        let model = QuillCodeWorkspaceModel()
-
-        model.setDraft("/")
-        var suggestions = model.surface().composer.slashSuggestions
-        XCTAssertEqual(suggestions.prefix(3).map(\.usage), ["/help", "/status", "/new"])
-
-        model.setDraft("/workt")
-        suggestions = model.surface().composer.slashSuggestions
-        XCTAssertEqual(suggestions.first?.usage, "/worktrees")
-        XCTAssertEqual(suggestions.first?.insertText, "/worktrees")
-
-        model.setDraft("/fol")
-        suggestions = model.surface().composer.slashSuggestions
-        XCTAssertEqual(suggestions.first?.usage, "/follow-up when")
-        XCTAssertEqual(suggestions.first?.insertText, "/follow-up in ")
-
-        model.setDraft("/workspace-c")
-        suggestions = model.surface().composer.slashSuggestions
-        XCTAssertEqual(suggestions.first?.usage, "/workspace-check when")
-        XCTAssertEqual(suggestions.first?.insertText, "/workspace-check in ")
-
-        model.setDraft("/project r")
-        suggestions = model.surface().composer.slashSuggestions
-        XCTAssertEqual(suggestions.prefix(2).map(\.usage), ["/project refresh", "/project rename name"])
-
-        model.setDraft("/pr l")
-        suggestions = model.surface().composer.slashSuggestions
-        XCTAssertEqual(suggestions.first?.usage, "/pr labels add|remove label")
-        XCTAssertEqual(suggestions.first?.insertText, "/pr labels add ")
-
-        model.setDraft("run /help")
-        XCTAssertEqual(model.surface().composer.slashSuggestions, [])
-    }
-
-    func testShortcutRegistryLabelsSurfaceCommands() {
-        let model = QuillCodeWorkspaceModel()
-        let commandsByID = Dictionary(uniqueKeysWithValues: model.surface().commands.map { ($0.id, $0) })
-
-        for shortcut in WorkspaceShortcutRegistry.shortcuts {
-            XCTAssertEqual(
-                commandsByID[shortcut.commandID]?.shortcut,
-                shortcut.displayLabel,
-                shortcut.commandID
-            )
-        }
-    }
-
-    func testShortcutRegistryHasNoDuplicateBindings() {
-        let bindings = WorkspaceShortcutRegistry.shortcuts.map {
-            "\($0.modifiers.map(\.rawValue).joined(separator: "+"))+\($0.key)"
-        }
-
-        XCTAssertEqual(Set(bindings).count, bindings.count)
-    }
-
-    func testCommandPaletteRanksByShortcutKeywordsAndTitle() {
-        let model = QuillCodeWorkspaceModel()
-        let commands = model.surface().commands
-
-        XCTAssertEqual(
-            WorkspaceCommandPalette.rankedCommands(commands, matching: "shell").first?.id,
-            "toggle-terminal"
-        )
-        XCTAssertEqual(
-            WorkspaceCommandPalette.rankedCommands(commands, matching: "cmd+k").first?.id,
-            "search"
-        )
-        XCTAssertEqual(
-            WorkspaceCommandPalette.rankedCommands(commands, matching: "cmd+f").first?.id,
-            "find-in-chat"
-        )
-        XCTAssertEqual(
-            WorkspaceCommandPalette.rankedCommands(commands, matching: "create pull").first?.id,
-            "git-pr-create"
-        )
-        XCTAssertEqual(
-            WorkspaceCommandPalette.rankedCommands(commands, matching: "checks").first?.id,
-            "git-pr-checks"
-        )
-        XCTAssertEqual(
-            WorkspaceCommandPalette.rankedCommands(commands, matching: "pr diff").first?.id,
-            "git-pr-diff"
-        )
-        XCTAssertEqual(
-            WorkspaceCommandPalette.rankedCommands(commands, matching: "checkout pull").first?.id,
-            "git-pr-checkout"
-        )
-        XCTAssertEqual(
-            WorkspaceCommandPalette.rankedCommands(commands, matching: "request reviewers").first?.id,
-            "git-pr-reviewers"
-        )
-        XCTAssertEqual(
-            WorkspaceCommandPalette.rankedCommands(commands, matching: "comment pull").first?.id,
-            "git-pr-comment"
-        )
-        XCTAssertEqual(
-            WorkspaceCommandPalette.rankedCommands(commands, matching: "approve pr").first?.id,
-            "git-pr-review"
-        )
-        XCTAssertEqual(
-            WorkspaceCommandPalette.rankedCommands(commands, matching: "label pr").first?.id,
-            "git-pr-labels"
-        )
-        XCTAssertEqual(
-            WorkspaceCommandPalette.rankedCommands(commands, matching: "merge pull").first?.id,
-            "git-pr-merge"
-        )
-        XCTAssertEqual(
-            WorkspaceCommandPalette.rankedCommands(commands, matching: "cmd+/").first?.id,
-            "keyboard-shortcuts"
-        )
-        XCTAssertEqual(
-            WorkspaceCommandPalette.rankedCommands(commands, matching: "shortcuts").first?.id,
-            "keyboard-shortcuts"
-        )
-        XCTAssertEqual(
-            WorkspaceCommandPalette.rankedCommands(commands, matching: ">shell").first?.id,
-            "toggle-terminal"
-        )
-        XCTAssertEqual(
-            WorkspaceCommandPalette.rankedCommands(commands, matching: "/mode").first?.title,
-            "/mode auto|review|read-only"
-        )
-        XCTAssertFalse(
-            WorkspaceCommandPalette.rankedCommands(commands, matching: "")
-                .contains { $0.id.hasPrefix(SlashCommandCatalog.commandPaletteIDPrefix) }
-        )
-        XCTAssertTrue(
-            WorkspaceCommandPalette.rankedCommands(commands, matching: "mode")
-                .contains { $0.id.hasPrefix(SlashCommandCatalog.commandPaletteIDPrefix) }
-        )
-        XCTAssertEqual(
-            WorkspaceCommandPalette.groupedCommands(commands, matching: "/").map(\.title),
-            [WorkspaceCommandPalette.slashCategory]
-        )
-    }
-
-    func testCommandPaletteGroupsFilteredResultsByCategory() {
-        let model = QuillCodeWorkspaceModel()
-        let groups = WorkspaceCommandPalette.groupedCommands(model.surface().commands, matching: ">worktree")
-
-        XCTAssertEqual(groups.map(\.title), [WorkspaceCommandPalette.gitCategory])
-        XCTAssertEqual(groups.first?.commands.map(\.id), [
-            "git-worktree-list",
-            "git-worktree-create",
-            "git-worktree-remove"
-        ])
-    }
-
-    func testWorkspaceCommandSurfaceDecodesOlderPayloadWithoutCategoryMetadata() throws {
-        let data = #"{"id":"search","title":"Search","shortcut":"Cmd+K","isEnabled":true}"#.data(using: .utf8)!
-
-        let command = try JSONDecoder().decode(WorkspaceCommandSurface.self, from: data)
-
-        XCTAssertEqual(command.category, WorkspaceCommandPalette.workspaceCategory)
-        XCTAssertEqual(command.keywords, [])
-    }
-
-    func testContextBannerDecodesOlderPayloadWithoutCompactCommand() throws {
-        let data = """
-        {
-          "usedPercent": 88,
-          "title": "Approaching context limit",
-          "subtitle": "Older turns may drop out soon.",
-          "newThreadCommand": {"id":"new-chat","title":"New thread"},
-          "forkCommand": {"id":"fork-from-last","title":"Fork from last","isEnabled":true}
-        }
-        """.data(using: .utf8)!
-
-        let banner = try JSONDecoder().decode(ContextBannerSurface.self, from: data)
-
-        XCTAssertEqual(banner.usedPercent, 88)
-        XCTAssertEqual(banner.compactCommand.id, "compact-context")
-        XCTAssertEqual(banner.compactCommand.title, "Compact context")
-        XCTAssertEqual(banner.compactCommand.category, WorkspaceCommandPalette.threadCategory)
-        XCTAssertEqual(banner.compactCommand.isEnabled, true)
-    }
-
     func testSurfaceIncludesLocalEnvironmentActionCommands() {
         let project = ProjectRef(
             name: "QuillCode",
@@ -504,72 +269,6 @@ final class WorkspaceSurfaceTests: XCTestCase {
         XCTAssertFalse(surface.review.isVisible)
         XCTAssertFalse(surface.composer.canSend)
         XCTAssertTrue(surface.topBar.showsComputerUseSetup)
-    }
-
-    func testContextBannerAppearsNearEstimatedLimit() throws {
-        let longMessage = "context " + String(repeating: "word ", count: 26_000)
-        let thread = ChatThread(title: "Long context", messages: [
-            .init(role: .user, content: longMessage)
-        ])
-        let model = QuillCodeWorkspaceModel(root: QuillCodeRootState(
-            threads: [thread],
-            selectedThreadID: thread.id
-        ))
-
-        let surface = model.surface()
-        let banner = try XCTUnwrap(surface.contextBanner)
-
-        XCTAssertTrue(banner.usedPercent >= 80)
-        XCTAssertTrue(banner.title.contains("Context"))
-        XCTAssertEqual(banner.newThreadCommand.id, "new-chat")
-        XCTAssertEqual(banner.forkCommand.id, "fork-from-last")
-        XCTAssertEqual(banner.compactCommand.id, "compact-context")
-        XCTAssertEqual(surface.commands.first { $0.id == "fork-from-last" }?.isEnabled, true)
-        XCTAssertEqual(surface.commands.first { $0.id == "compact-context" }?.isEnabled, true)
-    }
-
-    func testContextBannerHiddenForShortThreadAndForkDisabledWithoutMessages() {
-        let thread = ChatThread(title: "Short")
-        let model = QuillCodeWorkspaceModel(root: QuillCodeRootState(
-            threads: [thread],
-            selectedThreadID: thread.id
-        ))
-
-        let surface = model.surface()
-
-        XCTAssertNil(surface.contextBanner)
-        XCTAssertEqual(surface.commands.first { $0.id == "fork-from-last" }?.isEnabled, false)
-        XCTAssertEqual(surface.commands.first { $0.id == "compact-context" }?.isEnabled, false)
-    }
-
-    func testSidebarSearchFiltersByThreadTitleSubtitleAndTranscriptContent() {
-        let selectedThread = ChatThread(title: "Run whoami", model: TrustedRouterDefaults.synthModel)
-        var otherThread = ChatThread(title: "Review git diff", model: "z-ai/glm-5.2", isPinned: true)
-        otherThread.messages = [
-            .init(role: .user, content: "Can you inspect the browser preview?")
-        ]
-        var archivedThread = ChatThread(title: "Old release plan", model: TrustedRouterDefaults.synthModel)
-        archivedThread.isArchived = true
-        let surface = SidebarSurface(
-            items: [
-                SidebarItemSurface(item: SidebarItem(thread: selectedThread), selectedThreadID: selectedThread.id),
-                SidebarItemSurface(item: SidebarItem(thread: otherThread), selectedThreadID: selectedThread.id),
-                SidebarItemSurface(item: SidebarItem(thread: archivedThread), selectedThreadID: selectedThread.id)
-            ],
-            selectedThreadID: selectedThread.id
-        )
-
-        XCTAssertEqual(surface.filteredItems(matching: "").map(\.title), ["Run whoami", "Review git diff", "Old release plan"])
-        XCTAssertEqual(surface.filteredItems(matching: "who").map(\.title), ["Run whoami"])
-        XCTAssertEqual(surface.filteredItems(matching: "GLM").map(\.title), ["Review git diff"])
-        XCTAssertEqual(surface.filteredItems(matching: "browser preview").map(\.title), ["Review git diff"])
-        XCTAssertEqual(surface.filteredItems(matching: "archived").map(\.title), ["Old release plan"])
-        XCTAssertTrue(surface.filteredItems(matching: "workspace manager").isEmpty)
-        XCTAssertEqual(surface.pinnedItems.map(\.title), ["Review git diff"])
-        XCTAssertEqual(surface.recentItems.map(\.title), ["Run whoami"])
-        XCTAssertEqual(surface.recentSections().map(\.title), ["Today"])
-        XCTAssertEqual(surface.archivedItems.map(\.title), ["Old release plan"])
-        XCTAssertEqual(surface.archivedItems.first?.actions.map(\.kind), [.unarchive, .delete])
     }
 
 }
